add experimental entity_behavior component to enable custom behavior
this can also open possibilities for scripting entities; for now it
will be used to add custom behaviour to entities but still retaining
the ability to use components, for example BossEntity relies on
position and velocity components to be attached to the entity.
Missing components may generate obscure runtime errors.